requirements.txt
flask
numpy
Pillow
onnxruntime
pytest
requests